+++ apps/meteor/app/livechat/server/lib/Helper.ts
@@ -43,6 +43,7 @@ import { RoutingManager } from './RoutingManager';
 import { isVerifiedChannelInSource } from './contacts/isVerifiedChannelInSource';
 import { migrateVisitorIfMissingContact } from './contacts/migrateVisitorIfMissingContact';
 import { getOnlineAgents } from './getOnlineAgents';
+import { saveTransferHistory } from './transfer';
 import { callbacks } from '../../../../lib/callbacks';
 import { validateEmail as validatorFunc } from '../../../../lib/emailValidator';
 import { i18n } from '../../../../server/lib/i18n';
@@ -516,7 +517,7 @@ export const forwardRoomToAgent = async (room: IOmnichannelRoom, transferData: T
 		return false;
 	}
 
-	await LivechatTyped.saveTransferHistory(room, transferData);
+	await saveTransferHistory(room, transferData);
 
 	const { servedBy } = roomTaken;
 	if (servedBy) {
@@ -632,7 +633,7 @@ export const forwardRoomToDepartment = async (room: IOmnichannelRoom, guest: ILi
 		(department?.allowReceiveForwardOffline && !(await LivechatTyped.checkOnlineAgents(departmentId)))
 	) {
 		logger.debug(`Room ${room._id} will be on department queue`);
-		await LivechatTyped.saveTransferHistory(room, transferData);
+		await saveTransferHistory(room, transferData);
 		return RoutingManager.unassignAgent(inquiry, departmentId, true);
 	}
 
@@ -692,7 +693,7 @@ export const forwardRoomToDepartment = async (room: IOmnichannelRoom, guest: ILi
 		);
 	}
 
-	await LivechatTyped.saveTransferHistory(room, transferData);
+	await saveTransferHistory(room, transferData);
 	if (oldServedBy) {
 		// if chat is queued then we don't ignore the new servedBy agent bcs at this
 		// point the chat is not assigned to him/her and it is still in the queue
